# Speed up buffer creation.