Since 1d20f50f8b4c62779c0dbfb1fa401168ab52dce0 (#40), instances already gained all the `Buffer`'s
methods via  prototype re-assignment. Skipping the `_augment`
call yields significant speedups:

+ ~50% faster small buffer creation:

  + `new(16)` before x 595,757 ops/sec ±0.92% (91 runs sampled)
  + `new(16)` after x 894,462 ops/sec ±0.69% (87 runs sampled)

+ ~20% faster slice:

  + `slice` before: 2,176,194 ops/sec ±1.59% (89 runs sampled)
  + `slice` after: 2,624,209 ops/sec ±0.42% (99 runs sampled)

Other benchmarks aren't significantly different. (These numbers are for v8.)

Also removed the similarly redundant `_isBuffer` property. In case you'd
like to keep it (even though it's private), let me know.