@dcousens 2016-01-08
LGTM, but I think @feross would be the most knowledgeable about all the subtleties for this.