@mtth 2016-01-08
Thanks @dcousens !

@feross  - I'm also curious if you would consider adding an option to not rely on prototype mutation (falling back to "augmenting" array instances). Without it, buffers are still unreasonably slow in some (most?) engines. For example here are results for Firefox 43:

+ BrowserBuffer#bracket-notation x 487,929 ops/sec ±1.33% (59 runs sampled) # before
+ BrowserBuffer#bracket-notation x 494,476 ops/sec ±1.19% (66 runs sampled) # after
+  Uint8Array#bracket-notation x 24,571,591 ops/sec ±1.12% (65 runs sampled) # !

I initially thought that clients would just be able to set `window.TYPED_ARRAY_SUPPORT` to false in these cases, but it feels a bit like a hack (since these browsers do support it), and it's slower (especially for slices, which then require a copy).

If this sounds reasonable to you, I'll update this pull request with a tentative implementation.