@feross 2016-01-08
We can't get rid of the `_isBuffer` property since that's needed for `is-buffer` (https://github.com/feross/is-buffer) to work in older Safari. `is-buffer` is used by browserify whenever there's a `Buffer.isBuffer` call (very common) to save requiring the whole buffer module in browser bundles.

Removing `_augment` will break typedarray-to-buffer (https://github.com/feross/typedarray-to-buffer) but that can be fixed by moving `_augment` directly into that package.

> I'm also curious whether you would consider adding an option to not rely on prototype mutation (falling back to "augmenting" array instances)

I don't want to expose any options that aren't in the node buffer, because then we'll have code that's browser-specific and could end up doing something different in future versions of node. We should solve the issue directly in this module, but that should be done in another issue/PR, IMO.