@mtth 2016-01-08
> We can't get rid of the _isBuffer property since that's needed for is-buffer (https://github.com/feross/is-buffer) to work in older Safari.

Sounds good. I'll add the `_isBuffer` property back.

> Removing _augment will break typedarray-to-buffer (https://github.com/feross/typedarray-to-buffer) but that can be fixed by moving _augment directly into that package.

I'll restore `_augment` as well then. (It might also come in handy if some fallback mechanism is implemented.)

> I don't want to expose any options that aren't in the node buffer, because then we'll have code that's browser-specific and could end up doing something different in future versions of node. We should solve the issue directly in this module, but that should be done in another issue/PR, IMO.

I'm fine doing the detection inside this library (though there's the risk of being suboptimal for future browser implementations). Also, maybe I didn't explain myself correctly, I didn't mean an option at instantiation time but rather something like what `TYPED_ARRAY_SUPPORT` currently does.