@feross 2016-01-09
I can fix up the issues and merge. I'll add new tests to catch the issues I manually caught.