@feross 2016-01-09
We can actually just remove `Buffer._augment` entirely. I'll update `typedarray-to-buffer` to do just modify `__proto__` like package now does.