@feross 2016-01-09
Released as v4.1.0.