@feross 2016-01-09
This is released in browserify v13: https://github.com/substack/node-browserify/blob/master/changelog.markdown